updated to reflect latest persistentID schema
According to: https://wiki.shibboleth.net/confluence/display/IDP30/PersistentNameIDGenerationConfiguration , persistent ID schema now MUST use of a primary key tuple that will exceed 1000 characters in mysql in UTF8.  
As this table is critical to the storage of identifiers, it is critical to keep in alignment with the reference implementation.