resolved meow issue and insideous regex
cat cat -- aka the meow issue happened when cat was used twice when only one cat was needed.
Insideous regex was detected when sed would ingest a openssl rand base64'd string but a trailing slash borked it.
We now suppress a few interesting characters to construct the 'esalt' value we use.